Make frets u8, remove Partline position tracking
There is no reason to have support for frets > 256.
Partline tracking was there but wasn't actually used for any errors.

A 0-4% performance gain.

+++ src/backend/errors/backend_error.rs
@@ -53,7 +53,7 @@ impl<'a> BackendError<'a> {
         location_a: usize,
         location_b: usize,
         string_name: char,
-        fret: u16,
+        fret: u8,
         diagnostics: Vec<Diagnostic>,
     ) -> Self {
         Self {
